feat: migrate flower management from seller to admin
- Create AdminFlowerController with admin authorization instead of seller
- Implement AdminFlowerService to handle flower CRUD operations for admin
- Add AdminFlowerValidation with admin-specific validation logic
- Remove SellerId and SellerShopName from FlowerResponse DTO
- Remove SellerId from CreateFlowerRequest DTO
- Update Program.cs to register AdminFlowerService instead of seller services
- Remove seller-related includes and methods from Common FlowerService
- Delete SellerFlowerManagement controller and related services
- Admin can now create and manage all flowers in the system
- New API endpoints: /api/admin/flowers/* replacing /api/seller/flowers/*
- Flowers are no longer tied to specific sellers
