+++ PlatformFlower/Services/Common/Flower/IFlowerService.cs
@@ -7,6 +7,5 @@ namespace PlatformFlower.Services.Common.Flower
         Task<List<FlowerResponse>> GetActiveFlowersAsync();
         Task<FlowerResponse?> GetActiveFlowerByIdAsync(int flowerId);
         Task<List<FlowerResponse>> GetFlowersByCategoryAsync(int categoryId);
-        Task<List<FlowerResponse>> GetFlowersBySellerAsync(int sellerId);
     }
 }
