--- PlatformFlower/Services/Seller/FlowerManagement/IFlowerManagementService.cs
@@ -1,11 +1,0 @@
-using PlatformFlower.Models.DTOs.Flower;
-
-namespace PlatformFlower.Services.Seller.FlowerManagement
-{
-    public interface IFlowerManagementService
-    {
-        Task<FlowerResponse> ManageFlowerAsync(CreateFlowerRequest request, int? currentSellerId = null);
-        Task<List<FlowerResponse>> GetAllFlowersAsync();
-        Task<FlowerResponse?> GetFlowerByIdAsync(int flowerId);
-    }
-}
